// onto-demo-server/src/types.ts
export type NewCredentialProps = {
    firstName?: string
    lastName?: string
    email?: string
    department?: string
    securityClearanceLevel?: string
    phone?: string
    address?: string
    roomType?: string
    roomNumber?: string
    checkIn?: string
    checkOut?: string
    adminLoginType: string
}